bugfix 1등 여러개 에러 수정, LOTTO_REWARD 수정

+++ src/main/java/lotto/domain/result/LottoGameResult.java
@@ -24,7 +24,7 @@ public class LottoGameResult {
         lottoMatchResults.addMatchNumber(matchNumber);
     }
 
-    int getTotalReward() {
+    long getTotalReward() {
         return lottoMatchResults.getLottoReward();
     }
